refactor: validateTest file renamed
