+++ backend/src/utils/validation/courses/index.ts
@@ -3,7 +3,7 @@ import isValidComplexity from './isValidComplexity';
 import isValidCourseData from './isValidCourseData';
 import validateMaterials from './isValidMaterials';
 import isValidTechnologies from './isValidTechnologies';
-import validateTest from './isValidTest';
+import validateTest from './validateTest';
 import isValidText from './isValidText';
 import validateTitle from './validateTitle';
 import validatePlainMaterial from './validatePlainMaterial';
